Extract UpdateIndexCallable
Extract UpdateIndexCallable from IndexUpdater into a standalone class, bringing with
it methods only used by the callable in IndexUpdater, decoupling the callable from
the IndexUpdater.
