+++ cloudant-sync-datastore-core/src/main/java/com/cloudant/sync/internal/query/ValueExtractor.java
@@ -25,7 +25,7 @@ import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-class ValueExtractor {
+public class ValueExtractor {
 
     private static final Logger logger = Logger.getLogger(ValueExtractor.class.getName());
